+++ core/execution/engine_snapshot_test.go
@@ -202,8 +202,6 @@ func getSpotMarketConfig() *types.Market {
 				TimeWindow:    101,
 				ScalingFactor: num.DecimalFromFloat(1.0),
 			},
-			TriggeringRatio:  num.DecimalZero(),
-			AuctionExtension: 0,
 		},
 		Fees: &types.Fees{
 			Factors: &types.FeeFactors{
@@ -285,8 +283,6 @@ func getMarketConfig() *types.Market {
 				TimeWindow:    101,
 				ScalingFactor: num.DecimalFromFloat(1.0),
 			},
-			TriggeringRatio:  num.DecimalFromFloat(0.9),
-			AuctionExtension: 10000,
 		},
 		Fees: &types.Fees{
 			Factors: &types.FeeFactors{