+++ vegatools/checkpoint/type.go
@@ -263,8 +263,6 @@ func dummy() *all {
 								TimeWindow:    10,
 								ScalingFactor: 0.7,
 							},
-							TriggeringRatio:  "0.5",
-							AuctionExtension: 10,
 						},
 						RiskParameters: &vega.NewMarketConfiguration_LogNormal{
 							LogNormal: &vega.LogNormalRiskModel{